protected routes in goals

+++ backend/controllers/goalController.js
@@ -1,18 +1,16 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 //-----------GET GOALS-------
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
 //-----------SET GOALS VALUES-------
 const setGoal = asyncHandler(async (req, res) => {
-  const { text } = req.body;
-  console.log(text);
-
   if (!req.body.text) {
     res.status(400);
     throw new Error("Please Enter a text");
@@ -20,6 +18,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -31,6 +30,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     req.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //CHECK for the user
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -47,6 +60,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+  const user = await User.findById(req.user.id);
+
+  //CHECK for the user
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   await goal.deleteOne();
